+++ tailwind.config.js
@@ -1,5 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  mode: "jit",
-  content: ["./internal/transport/http/views/**/*.templ"],
-};